refactor: extract EmailSettingsForm instead of gating email in shared form
ChangeUserSettingsForm carried all the email state and handlers even in
the helper flow, where they could not render, and forced callers to pass
a meaningless email="" plus a showEmailSettings flag. Move the email
settings into their own EmailSettingsForm owned by ProfileView;
ChangeUserSettingsForm is now password-only and drops both props.
