+++ client/src/views/profile/ProfileView.tsx
@@ -3,6 +3,7 @@ import styled from "styled-components";
 import { useTranslation } from "react-i18next";
 import { useNavigate } from "react-router";
 import { ChangeUserSettingsForm } from "client/views/helper/components/ChangeUserSettingsForm";
+import { EmailSettingsForm } from "client/views/profile/components/EmailSettingsForm";
 import { useAppSelector } from "client/utils/hooks";
 import { Button, ButtonStyle } from "client/components/Button";
 import { AppRoute } from "client/app/AppRoutes";
@@ -36,11 +37,10 @@ export const ProfileView = (): ReactElement => {
           {t("button.logout")}
         </StyledButton>
       </UserInfoContainer>
+      <EmailSettingsForm email={email} />
       <ChangeUserSettingsForm
         isLocalLogin={isLocalLogin}
-        email={email}
         usernameToUpdate={username}
-        showEmailSettings={true}
       />
     </Container>
   );
